pyproject.toml
[tool.poetry]
name = "nonebot-plugin-weather-rank"
version = "0.1.8"
description = "一个比较群友地区天气温度的插件"
authors = ["hanasa2023 <hanasakayui2022@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9, <4.0"
httpx = ">=0.27.0"
nonebot2 = ">=2.3.2"
pydantic = ">=1.0, <3.0.0"
beautifulsoup4 = ">=4.12.3"
nonebot-adapter-onebot = ">=2.1.0"
nonebot-plugin-alconna = ">=0.52.2"
nonebot-plugin-htmlrender = ">=0.3.5"
nonebot-plugin-apscheduler = "^0.5.0"
tortoise-orm = { extras = ["asyncpg"], version = "^0.21.6" }
nonebot-plugin-localstore = "^0.7.1"

[[tool.poetry.source]]
name = "mirrors"
url = "https://mirrors.tuna.tsinghua.edu.cn/pypi/web/simple/"
priority = "primary"

[tool.poetry.group.test.dependencies]
nonebug = "^0.3.7"
pytest-asyncio = "^0.24.0"

[tool.poetry.group.dev.dependencies]
mypy = ">=1.11.1"
nonebot2 = { version = ">=2.3.2", extras = ["fastapi", "uvicorn"] }
nonebot-adapter-onebot = ">=2.1.0"
ruff = ">=0.6.1"
types-beautifulsoup4 = ">=4.12.0.20240511"
pre-commit = ">=3.8.0"
basedpyright = "^1.18.2"

[tool.basedpyright]
typeCheckingMode = "standard"

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.nonebot]
adapters = [
  { name = "OneBot V11", module_name = "nonebot.adapters.onebot.v11" },
]
plugins = []
plugin_dirs = [""]
builtin_plugins = []
